+++ sdgx/models/ml/single_table/new_ctgan.py
@@ -1,0 +1,11 @@
+from sdgx.models.extension import hookimpl
+from sdgx.models.ml.single_table.base import MLSynthesizerModel
+
+
+class CTGANSynthesizerModel(MLSynthesizerModel):
+    pass
+
+
+# @hookimpl
+# def register(manager):
+#     manager.register("CTGAN", CTGANSynthesizerModel)
